+++ src/Cwaff-Guns/SeltzerPelter.cs
@@ -15,7 +15,7 @@ public class SeltzerPelter : CwaffGun
     public static void Init()
     {
         Gun gun = Lazy.SetupGun<SeltzerPelter>(ItemName, ShortDescription, LongDescription, Lore)
-          .SetAttributes(quality: ItemQuality.B, gunClass: GunClass.CHARGE, reloadTime: 1.0f, ammo: 150, shootFps: 36, muzzleFrom: Items.Mailbox)
+          .SetAttributes(quality: ItemQuality.B, gunClass: GunClass.RIFLE, reloadTime: 1.0f, ammo: 150, shootFps: 36, muzzleFrom: Items.Mailbox)
           .AddToShop(ItemBuilder.ShopType.Goopton);
 
         _ReloadAnimations = new(){